bloodlink-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose
passlib[bcrypt]